+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Allow `BTreeMap` to be initialized with a `Record<string, any>` object (in addition to `Map`)
 - Allow for `HashMap<KeyType, ValueType>` definitions
 - Refresh the `NetworkState` types as returned by the `rpc.system.networkState()` call to current
+- Expose `registry.createClass(<typeName>)`, to allow for creating type classes. If you are using `createClass` or `ClassOf` in your code, it is recommended to swap to this instance.
 
 # 1.4.1 Feb 26, 2020